Fix All functions

+++ src/modules/handlers/handlers.js
@@ -1,18 +1,16 @@
 "use strict";
 
-import { categoryModel } from "../../../database/models/category.model.js";
+import slugify from "slugify";
 import { catchError } from "../../middleware/catchError.js";
 import { ApiFeatures } from "../../utils/apiFeatures.js";
 import { AppError } from "../../utils/appError.js";
 
+//! add task not found
 const addOne = (model) => {
   return catchError(async (req, res, next) => {
     try {
       if (req.user) { req.body.createdBy = req.user._id; }
-      if (req.body.category) {
-        const category = await categoryModel.findById(req.body.category);
-        if (!category) return next(new AppError("Category not found", 404));
-      }
+      if (req.body.name) req.body.slug = slugify(req.body.name);
       const document = new model(req.body);
       await document.save();
       res.status(201).json({ message: "success", document });
@@ -39,11 +37,6 @@ const findOne = (model) => {
   return catchError(async (req, res, next) => {
     try {
       const document = await model.findById(req.params.id);
-      // To check if anyone is trying to get a private task by typing a query into the URL like => ?visible=private
-      if (document.visible == "private")
-        return next(
-          new AppError("You don't have permission to perform this action.", 403)
-        );
       !document && next(new AppError("Document not found.", 404));
       document && res.status(200).json({ message: "success", document });
     } catch (e) {
@@ -52,22 +45,12 @@ const findOne = (model) => {
   });
 };
 
+//! add task not found
 const updateOne = (model) => {
   return catchError(async (req, res, next) => {
     try {
-      let document;
-      if (req.user) {
-        document = await model.findOneAndUpdate(
-          {
-            _id: req.params.id,
-            $or: [{ createdBy: req.user._id }, { user: req.user._id }],
-          },
-          req.body,
-          { new: true }
-        );
-      } else {
-        document = await model.findByIdAndUpdate(req.params.id, req.body, { new: true, });
-      }
+      if (req.body.name) req.body.slug = slugify(req.body.name);
+      let document = await model.findOneAndUpdate({ _id: req.params.id, createdBy: req.user._id }, req.body, { new: true });
       !document && next(new AppError("Document not found.", 404));
       document && res.status(200).json({ message: "success", document });
     } catch (e) {
@@ -79,19 +62,7 @@ const updateOne = (model) => {
 const deleteOne = (model) => {
   return catchError(async (req, res, next) => {
     try {
-      let document;
-      if (req.user) {
-        document = await model.findOneAndDelete(
-          {
-            _id: req.params.id,
-            $or: [{ createdBy: req.user._id }, { user: req.user._id }],
-          },
-          req.body,
-          { new: true }
-        );
-      } else {
-        document = await model.findByIdAndDelete(req.params.id);
-      }
+      let document = await model.findOneAndDelete({ _id: req.params.id, createdBy: req.user._id, }, req.body, { new: true });
       !document && next(new AppError("Document not found.", 404));
       document && res.status(200).json({ message: "success" });
     } catch (e) {
